+++ clang/lib/CodeGen/CGBuiltin.cpp
@@ -5405,31 +5405,10 @@ RValue CodeGenFunction::EmitBuiltinExpr(const GlobalDecl GD, unsigned BuiltinID,
       return RValue::get(Builder.CreateZExt(Result, Int64Ty, "extend.zext"));
   }
   case Builtin::BI__builtin_setjmp: {
-    // Buffer is a void**.
     Address Buf = EmitPointerWithAlignment(E->getArg(0));
 
-    if (getTarget().getTriple().getArch() == llvm::Triple::systemz) {
-      // On this target, the back end fills in the context buffer completely.
-      // It doesn't really matter if the frontend stores to the buffer before
-      // calling setjmp, the back-end is going to overwrite them anyway.
-      Function *F = CGM.getIntrinsic(Intrinsic::eh_sjlj_setjmp);
-      return RValue::get(Builder.CreateCall(F, Buf.emitRawPointer(*this)));
-    }
-
-    // Store the frame pointer to the setjmp buffer.
-    Value *FrameAddr = Builder.CreateCall(
-        CGM.getIntrinsic(Intrinsic::frameaddress, AllocaInt8PtrTy),
-        ConstantInt::get(Int32Ty, 0));
-    Builder.CreateStore(FrameAddr, Buf);
-
-    // Store the stack pointer to the setjmp buffer.
-    Value *StackAddr = Builder.CreateStackSave();
-    assert(Buf.emitRawPointer(*this)->getType() == StackAddr->getType());
-
-    Address StackSaveSlot = Builder.CreateConstInBoundsGEP(Buf, 2);
-    Builder.CreateStore(StackAddr, StackSaveSlot);
-
-    // Call LLVM's EH setjmp, which is lightweight.
+    // The backend handles all buffer stores (FP, SP, IP) via
+    // @llvm.eh.sjlj.setjmp.
     Function *F = CGM.getIntrinsic(Intrinsic::eh_sjlj_setjmp);
     return RValue::get(Builder.CreateCall(F, Buf.emitRawPointer(*this)));
   }